pyproject.toml
[tool.poetry]
name = "epidemic-api"
version = "0.1.0"
description = ""
authors = ["Scrooge <76687908+SkrudjReal@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
redis = "5.0.4"
telethon = "^1.36.0"
pyrogram = "^2.0.106"
pydantic-settings = "^2.3.4"
tgcrypto = "^1.2.5"
pyjwt = "^2.8.0"
dependency-injector = "^4.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
